internx_backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
Pillow==11.1.0
reportlab==4.2.5
qrcode==8.0
google-api-python-client==2.120.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gunicorn==22.0.0
whitenoise==6.8.2
django-filter==24.3
drf-spectacular==0.27.2
